Segunda etapa: Melhorias implementadas no código do back que acessa o Banco de Dados.
