+++ back/.env
@@ -1,3 +1,3 @@
-DB_LOCAL="/src/mensagens.db"
+DB_PATH="/src/mensagens.db"
 URL_HOST="localhost"
 PORT=3000
