+++ back/src/index.ts
@@ -7,12 +7,13 @@ import * as WebSocket from "ws";
 import cors from "cors";
 import sqlite3 from 'sqlite3'
 import { open } from 'sqlite'
+import { getGroupIcon, setGroupIcon, getGroupName, setGroupName } from './acessDB';
 
 const app = express();
 export const server = http.createServer(app);
 export const wss = new WebSocket.Server({ server });
 
-const broadcast = (msg: any) => {
+export const broadcast = (msg: any) => {
   wss.clients.forEach((client) => {
     client.send(JSON.stringify(msg));
   });
@@ -102,97 +103,12 @@ async function insertMessage(stringfiedMessage: string) {
   })
 }
 
-//Funções para manipular o nome e o ícone do grupo
-app.get("/icon", (req: Request, res: Response) => {
-  async function getGroupIcon() {
-    const dbPath = path.resolve(__dirname, 'mensagens.db');
-    open({
-      filename: dbPath,
-      driver: sqlite3.Database
-    }).then((db) => {
-      db.all('SELECT icone FROM dadosGrupo WHERE id = 1').then((row) => {
-        res.json({ groupIcon: row[0].icone });
-      })
-    }).catch((err) => {
-      console.log("Erro ao abrir o DB: " + err)
-    })
-  }
-  getGroupIcon()
-});
-
-app.post("/icon", (req: Request, res: Response) => {
-  const body = req.body;
-  const groupIcon = body.groupIcon;
-  
-  async function updateGroupIcon() {
-    const dbPath = path.resolve(__dirname, 'mensagens.db');
-    open({
-      filename: dbPath,
-      driver: sqlite3.Database
-    }).then((db) => {
-      db.run('UPDATE dadosGrupo SET icone = (:icon) WHERE id = 1', {
-        ':icon': groupIcon
-      })
-    }).catch((err) => {
-      console.log("Erro ao abrir o DB: " + err)
-    })
-  }
-  updateGroupIcon()
-
-  broadcast({
-    type: "group",
-    groupIcon,
-  });
-});
-
-app.get("/group", (req: Request, res: Response) => {
-  async function getGroupName() {
-    const dbPath = path.resolve(__dirname, 'mensagens.db');
-    open({
-      filename: dbPath,
-      driver: sqlite3.Database
-    }).then((db) => {
-      db.all('SELECT nome FROM dadosGrupo WHERE id = 1').then((row) => {
-        res.json({ groupName: row[0].nome });
-      })
-    }).catch((err) => {
-      console.log("Erro ao abrir o DB: " + err)
-    })
-  }
-
-  getGroupName()
-});
-
-app.post("/group", (req: Request, res: Response) => {
-  const body = req.body;
-  const groupName = body.groupName;
-  
-  async function updateGroupName() {
-    const dbPath = path.resolve(__dirname, 'mensagens.db');
-    open({
-      filename: dbPath,
-      driver: sqlite3.Database
-    }).then((db) => {
-      db.run('UPDATE dadosGrupo SET nome = (:name) WHERE id = 1', {
-        ':name': groupName
-      })
-    }).catch((err) => {
-      console.log("Erro ao abrir o DB: " + err)
-    })
-  }
-  updateGroupName()
-
-  broadcast({
-    type: "group",
-    groupName,
-  });
-
-  res.json({ groupName });
-});
-
-
-
-
+//Segunda etapa: Funções para manipular o nome e o ícone do grupo no DB
+// agora serão importadas de outro arquivo
+app.get("/icon", getGroupIcon);
+app.post("/icon", setGroupIcon);
+app.get("/group", getGroupName);
+app.post("/group", setGroupName);
 
 app.post("/message", (req: Request, res: Response) => {
